[UPDATE] #product update table update complete

+++ views/product-index.php
@@ -24,14 +24,14 @@ $product = new Product($db);
 if($_POST) {
     include_once "../config/FileManager.php";
     $product->id = (int)$_POST["id"];
-    if(!$product->setProduct()) die('<label class="alert alert-danger width-100-percent" role="alert"> No product found. </label>');
+    if(!$product->setProduct()) die(json_encode(["response_html" => '<label class="alert alert-danger width-100-percent" role="alert"> No product found. </label>']));
     $file = new FileManager();
     if(file_exists($_FILES['feature_image']['tmp_name']) && is_uploaded_file($_FILES['feature_image']['tmp_name'])) {
         $old_image = $product->feature_image;
         $product->feature_image = $file->uploadImageAndGetPath('feature_image', Statics::getProductFeatureImageFullPath());
-        $file->deleteFile($old_image);
+        if($product->feature_image) $file->deleteFile($old_image);
     }
-    if(!$product->feature_image) die('<label class="alert alert-danger width-100-percent" role="alert"> An error occurred while uploading the file </label>');
+    if(!$product->feature_image) die(json_encode(["response_html" => '<label class="alert alert-danger width-100-percent" role="alert"> An error occurred while uploading the file </label>']));
 
     $product->product_title = $_POST['product_title'];
     $product->price = $_POST['product_price'];
@@ -39,13 +39,16 @@ if($_POST) {
     if(file_exists($_FILES['secondary_image']['tmp_name']) && is_uploaded_file($_FILES['secondary_image']['tmp_name'])) {
         $old_image = $product->secondary_image;
         $product->secondary_image = $file->uploadImageAndGetPath('secondary_image', Statics::getProductSecondaryImagePath());
-        if(!$product->secondary_image) die('<label class="alert alert-danger width-100-percent" role="alert"> An error occurred while uploading the secondary image file </label>');
+        if(!$product->secondary_image) die(json_encode(["response_html" => '<label class="alert alert-danger width-100-percent" role="alert"> An error occurred while uploading the secondary image file </label>']));
         $file->deleteFile($old_image);
     }
 
-    if($product->update())
-        die('<label class="alert alert-success width-100-percent" role="alert"> Success ! Product updated.</label>');
-    else die('<label class="alert alert-danger width-100-percent" role="alert"> Something went wrong. Try again.'.  $product->getErrors(). '</label>');
+    if($product->update()) {
+        die(json_encode(["response_html" => '<label class="alert alert-success width-100-percent" role="alert"> Success ! Product updated.</label>',
+            "code" => 200,
+            "updated_row" => $product->updatedTableRow()]));
+    }
+    else die(json_encode(["response_html" => '<label class="alert alert-danger width-100-percent" role="alert"> Something went wrong. Try again.'.  $product->getErrors(). '</label>']));
 }
 
 else {
@@ -92,15 +95,20 @@ include_once "layout/layout_foot.php";
 
         $("form#product_update").submit(function (e) {
             e.preventDefault();
+            let id = $('#product_id').val();
             let formData = new FormData(this);
             $.ajax({
                 url: window.location.pathname,
                 method: 'POST',
                 data: formData,
                 success: function (response) {
-                    $('#post_message_box').html(response);
+                    console.log(response);
+                    response = (JSON.parse(response));
+                    $('#post_message_box').html(response.response_html);
                     $(":input").val('');
                     $('#editProductModal').modal('hide');
+                    if (typeof response.code !== 'undefined' && response.code === 200)
+                        $('#'+id).replaceWith(response.updated_row);
                 },
                 cache: false,
                 contentType: false,
